Convert BossPhaseEvents into hook
